Split the distro_profile_system_test.py into three files
